requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
Pillow==9.3.0
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
pytest-django==4.5.2
django-extensions==3.2.3
